requirements.txt
boto3>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0 